+++ src/SFSI.h
@@ -14,6 +14,8 @@
 # define FCONE
 #endif
 
+void daxpy_set(int n, double *a, double *dx, int *ix, double *dy, int *iy);
+
 double dsum(int n, double *dx);
 
 double ddot3(int n, double *dx1, double *dx2, double *dx3);
@@ -41,25 +43,24 @@ void matrix_vector_product(int nrow,
 //                           double *B, double *C)
 
 void crossproduct(int nrow, int ncolA, int ncolB,
-                 double *A, double *B, double *C);
+                  double *A, double *B, double *C);
 
 void tcrossproduct(int nrowA, int ncol, int nrowB,
-                  double *A, double *B, double *C);
+                   double *A, double *B, double *C);
 
 void tcrossproduct_tri(int n, double *A, double *B, double *C);
 
 void crossproduct_scale(int nrow, int ncolA, int ncolB,
-                  double *A, double *d, double *B,
-                  double *C, double *work);
+                        double *A, double *d, double *B,
+                        double *C, double *work);
 
 void tcrossproduct_scale(int nrowA, int ncol, int nrowB,
-                   double *A, double *d, double *B,
-                   double *C, double *work);
+                         double *A, double *d, double *B,
+                         double *C, double *work);
 
 void slice_matrix(int nrow, double *A, double *x,
-                 int n, int *index,
-                 int k,
-                 int margin);
+                  int n, int *index,
+                  int k, int margin);
 
 void resize_matrix(int nrow, int ncol, int nrownew, int ncolnew, double *A);
 
@@ -75,19 +76,19 @@ void invert_matrix(int n, double *A, double *Ainv,
                    double *eps, double *work, int *info);
 
 void matrix_vector_product_subset(int nrow,
-                                 int ncol,
-                                 double *A,
-                                 double *x,
-                                 double *y,
-                                 int nirow, int *irow,
-                                 int nicol, int *icol,
-                                 int transpose,
-                                 double *work);
+                                  int ncol,
+                                  double *A,
+                                  double *x,
+                                  double *y,
+                                  int nirow, int *irow,
+                                  int nicol, int *icol,
+                                  int transpose,
+                                  double *work);
 
 void append_to_sorted_vector_integer(int n,
-                                    int *v,
-                                    int k,
-                                    int *values);
+                                     int *v,
+                                     int k,
+                                     int *values);
 
 void reduce_vector_integer(int n,
                            int *v,