+++ NEWS.md
@@ -1,3 +1,10 @@
+### Version 1.4.1 (Aug-23-2024)
+
+  - Some warnings were fixed in the documentation as required by CRAN.
+  - Fixed bugs in functions 'fitBLUP', 'SGP', 'getGenCov': an error was produced if 'y' has 2 dimensions but it is a 'data.frame'. This was fixed by using 'as.matrix(y)'
+  - Fixed bug in 'multitrait.plot' function
+  - Fixed bug in checkpoint in function 'fitBLUP': an error was produced whenever Z = NULL & K = NULL & ntraits > 1. This error is not produced if an EVD is provided
+
 ### Version 1.4 (Jun-19-2024)
 
   - Changes in function names: SSI -> SGP,  SSI.CV -> SGP.CV. Results from both functions are of the class "SGP" standing for "sparse genomic prediction"